+++ PathFinding/Models/Node.cs
@@ -5,7 +5,7 @@
         public Position CurrentPos;
         public Position PreviousPos;
 
-        public GridNode(int x, int y)
+        public GridNode(short x, short y)
         {
             CurrentPos = new Position(x, y);
             PreviousPos = new Position(x, y);
